-- Changes to get the code to compile with gcc 4.3  -- 32-bit support: replace size_t by off_t for file sizes     so that it works properly on 32-bit platforms with     large file support enabled

+++ src/cc/chunk/KfsOps.cc
@@ -509,7 +509,7 @@ int
 ReadOp::HandleDone(int code, void *data)
 {
     IOBuffer *b;
-    size_t chunkSize = 0;
+    off_t chunkSize = 0;
 
     if (code == EVENT_DISK_ERROR)
         status = -1;
